Fix bugs with the conversion from fgetln to getline
The length set by fgetln is always the length of the line ending at the
newline character, while the length set by getline includes the null
byte and any extra bytes that may be allocated for the buffer. Change
the getline calls that were converted from fgetln to use the return
value of getline in place of the buffer length parameter.

Update the example code in DIFFERENCES to separate the line length and
buffer length variables.